apps/api/app/core/config.py
import os
from dotenv import load_dotenv

load_dotenv()

CORS_ORIGINS = os.getenv("CORS_ORIGINS", "http://localhost:3000").split(",")
GITHUB_TOKEN = os.getenv("GITHUB_TOKEN", "")
GITHUB_USERNAME = os.getenv("GITHUB_USERNAME", "souzxxx")
METRICS_INTERVAL = float(os.getenv("METRICS_INTERVAL", "2.0"))
