apps/api/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==15.0.1
psutil==7.0.0
python-dotenv==1.1.0
httpx==0.28.1
pytest==8.4.1
pytest-asyncio==0.26.0
